+++ src/cpu.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-vector<int> runCPU(vector<float> &B, vector<float> &C, vector<float> &values, vector<int> &rowIndices, vector<int> &columnIndices, vector<int> &updatedVertexIndices, int numberOfRows, int threadCount) {
+vector<int> runCPU(vector<int> &B, vector<int> &C, vector<int> &values, vector<int> &rowIndices, vector<int> &columnIndices, vector<int> &updatedVertexIndices, int numberOfRows, int threadCount) {
     vector<int> newUpdatedVertexIndices;
 
     #pragma omp parallel for num_threads(threadCount)
@@ -12,7 +12,7 @@ vector<int> runCPU(vector<float> &B, vector<float> &C, vector<float> &values, ve
         int source = updatedVertexIndices[i];
 		for (int j = rowIndices[source]; j < (source == numberOfRows - 1 ? columnIndices.size() : rowIndices[source + 1]); j++) {
 			int target = columnIndices[j];
-			float value = values[j];
+			int value = values[j];
 			if (B[source] + value < C[target]) {
 				C[target] = B[source] + value;
 				newUpdatedVertexIndices.push_back(target);
